Xaan addition, Acrobat rework, and new script format
I finally decided to stop being lazy and redid the JSON input from scripts to work with the current format that TPI uses for their scripts on the app. Updated the input validation to handle this.

Fixed an encoding error that I only ran across when I started developing in PyCharm instead of repl.it

Added the Xaan to the night.json file

Updated the Acrobat to have the correct Other Night text.